Scraper/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
motor==3.3.2
pymongo==4.6.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
python-dateutil==2.8.2
